solution/worker/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.9
requests==2.32.3
python-docx==1.1.2
pdfminer.six==20240706
tenacity==9.0.0
